Package lifecycle (#698)
