+++ packages/agreement-lifecycle/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "pagopa-interop-agreement-lifecycle",
+  "version": "1.0.0",
+  "private": true,
+  "description": "PagoPA Interoperability agreement lifecycle",
+  "main": "dist",
+  "type": "module",
+  "exports": {
+    ".": "./dist/index.js"
+  },
+  "scripts": {
+    "lint": "eslint . --ext .ts,.tsx",
+    "lint:autofix": "eslint . --ext .ts,.tsx --fix",
+    "format:check": "prettier --check src",
+    "format:write": "prettier --write src",
+    "start": "node --loader ts-node/esm --watch ./src/index.ts",
+    "build": "tsc",
+    "check": "tsc --project tsconfig.check.json"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "Apache-2.0",
+  "dependencies": {
+    "@types/node": "20.14.6",
+    "pagopa-interop-models": "workspace:*"
+  },
+  "devDependencies": {
+    "eslint": "8.57.0",
+    "prettier": "2.8.8",
+    "typescript": "5.4.5"
+  }
+}
