+++ packages/agreement-process/package.json
@@ -46,6 +46,7 @@
     "mongodb": "6.7.0",
     "pagopa-interop-commons": "workspace:*",
     "pagopa-interop-models": "workspace:*",
+    "pagopa-interop-agreement-lifecycle": "workspace:*",
     "pagopa-interop-selfcare-v2-client": "workspace:*",
     "ts-pattern": "5.2.0",
     "uuid": "10.0.0",
